Implemented interruptable trajectory
The current implementation seems to work properly in simulation
but on real robot it is not working properly, i.e., the robot
is jerkingly moving. The previously created single file is split
into two files where the GUI publishes the stop/resume command
and the other node implements the interruptable trajectory logic.

The cancel topic was not found in the sim env but was found in the
hardward env. Using cancel topic seems like the best way to go ahead.
Tomorrow must work on this.